Add files via upload
Allows for the creation of types of CFG's, the usage being used in the code at the moment is for non-palindromes and A^nB^n, but it can also have other use case CFG's. The only nessecary thing is that I believe my class only allows for a single state to be passed in with all rules contained within. Could be an easy fix though.